ajustments

+++ 6.HTTP/include/timeout.h
@@ -7,9 +7,9 @@ namespace http {
 using timeout_t    = std::chrono::seconds;
 using time_point_t = std::chrono::system_clock::time_point;
 
-inline constexpr int ACCEPT_TIMEOUT = 5;
-inline constexpr int CONNECTION_TIMEOUT = 20;
-inline constexpr int KEEP_ALIVE_CONNECTION_TIMEOUT = 60;
+inline constexpr size_t ACCEPT_TIMEOUT = 5u;
+inline constexpr size_t CONNECTION_TIMEOUT = 20u;
+inline constexpr size_t KEEP_ALIVE_CONNECTION_TIMEOUT = 60u;
 
 }  // namespace http
